+++ shiva/shiva/helpers/misc.py
@@ -42,4 +42,4 @@ def warn_with_traceback(message, category, filename, lineno, file=None, line=Non
 
 def one_hot_from_logits(logits):
     # print(logits.max(1, keepdim=True))
-    return (logits == logits.max(1, keepdim=True)[0]).float()
+    return (logits == logits.max(1, keepdim=True)[0]).double()
