+++ shiva/shiva/helpers/networks_handler.py
@@ -29,27 +29,27 @@ def DynamicLinearSequential(input_dim, output_dim, layers: list, activ_function:
 
     net_layers = []
 
-    net_layers.append(nn.Linear(input_dim, layers[0]))
-    net_layers.append(activ_function[0]())
+    net_layers.append(nn.Linear(input_dim, layers[0]).double())
+    net_layers.append(activ_function[0]().double())
     in_layer_size = layers[0]
 
     if len(activ_function) > 1:
         for i, (out_layer_size, af) in enumerate(zip(layers, activ_function)):
             if i == 0: continue
-            net_layers.append(nn.Linear(in_layer_size, out_layer_size))
-            net_layers.append(af())
+            net_layers.append(nn.Linear(in_layer_size, out_layer_size).double())
+            net_layers.append(af().double())
             in_layer_size = out_layer_size
     else:
         af = activ_function[0]
         for i, out_layer_size in enumerate(layers):
             if i == 0: continue
-            net_layers.append(nn.Linear(in_layer_size, out_layer_size))
-            net_layers.append(af())
+            net_layers.append(nn.Linear(in_layer_size, out_layer_size).double())
+            net_layers.append(af().double())
             in_layer_size = out_layer_size
 
     if last_layer:
         last_layer_dim = layers[len(layers)-1] if len(layers) > 1 else layers[0]
-        net_layers.append( nn.Linear(last_layer_dim, output_dim) )
+        net_layers.append(nn.Linear(last_layer_dim, output_dim).double())
 
 
     for i, layer in enumerate(net_layers):
@@ -62,9 +62,9 @@ def DynamicLinearSequential(input_dim, output_dim, layers: list, activ_function:
 
     if output_function is not None:
         try:
-            net_layers.append(output_function(dim=-1))
+            net_layers.append(output_function(dim=-1).double())
         except TypeError:
-            net_layers.append(output_function())
+            net_layers.append(output_function().double())
 
     device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
-    return nn.Sequential(*net_layers).to(device)
+    return nn.Sequential(*net_layers).to(device).double()
